Add RS namespace to vendor section
There was an issue which vendor process has failed to load
libRS_internal because libft is private library. As we cannot allow
access to the private library from all vndk libs, we should add rs
namespace and narrow use cases. This change adds rs namespace to the
vendor section, and links to the namespace only with libRS_internal
library.

Bug: 199044450
Test: N/A
Change-Id: I6b6182831e32409b562e763090e73e16aa5cd54a